Add files via upload
An Innovative Product With Societal Acceptance Is The One That Aids The Comfort, Convenience And Efficiency In Everyday Life. Purchasing And Shopping At Big Malls Is Becoming Daily Activity In Metro Cities. We Can See Big Rush At These Malls On Holidays And Weekends. People Purchase Different Items And Put Them In Trolley. After Completion Of Purchases, One Needs To Go To Billing Counter For Payments. At Billing Counter The Cashier Prepare The Bill Using Bar Code Reader Which Is Very Time Consuming Process And Results In Long Queue At Billing Counter. In This Paper, We Discuss A Product "Smart Trolley In Mega Mall" Being Developed To Assist A Person In Everyday Shopping In Terms Of Reduced Time Spent While Purchasing. The Main Objective Of Proposed System Is To Provide A Technology Oriented, Low-Cost, Easily Scalable, And Rugged System For Assisting Shopping In Person